arrange routing & add /store/groups (this returns mock)
