+++ store/__init__.py
@@ -1,0 +1,4 @@
+from app import app
+
+import views
+
